telegram_bot/requirements.txt
aiogram==2.10.1
aiohttp==3.6.3
python-dotenv==0.14.0
requests==2.22.0
